refactor Image
